% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse fastq.R
\name{align_fastq_to_vdjca}
\alias{align_fastq_to_vdjca}
\title{align_fastq_to_vdjca}
\usage{
align_fastq_to_vdjca(java_param = "java -Xmx4g -Xms3g -jar", path_mixcr,
  input_dir, input_filename, report_name = "alignmentReport",
  replace = FALSE, string_only = FALSE)
}
\arguments{
\item{java_param}{a character string containing java parameters}

\item{path_mixcr}{the path to mixcr.jar}

\item{input_dir}{where the files are}

\item{input_filename}{file name (type R1), without the extension}

\item{report_name}{the name of the file alignmentReport. (without the extension)
Set to NULL to not create the file.}

\item{replace}{if TRUE, if the output file already exists, it will be replaced.}

\item{string_only}{if TRUE, don't execute the command but returns it in a string}
}
\value{
run command line in console. Creates .vdjca files
        in the same directory than the fastq.gz files
}
\description{
execute the command : 'align -OvParameters.geneFeatureToAlign=VTranscript
}
\author{
DEBOT Damien <damien.debot@selor.be>
}
