% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse fastq.R
\name{analyse_tcells}
\alias{analyse_tcells}
\title{analyse_tcells}
\usage{
analyse_tcells(java_param = "java -Xmx4g -Xms3g -jar", path_vdjtools,
  path_mixcr, input_dir, tcell_type = "auto", replace = FALSE,
  string_only = FALSE, ...)
}
\arguments{
\item{java_param}{a character string containing java parameters}

\item{path_vdjtools}{the path to vdjtools.jar}

\item{path_mixcr}{the path to mixcr.jar}

\item{input_dir}{directory where the files are (also look in sub-directories)}

\item{tcell_type}{The type of Tcells to analyse (A, B, G or D).
"auto" to let R detect the type of Tcells automatically.}

\item{replace}{if TRUE, if the output file already exists, it will be replaced.}

\item{string_only}{if TRUE, don't execute the command but returns it in a string}

\item{...}{Parameters to pass to the called functions (e.g. report_name)}
}
\value{
runs command lines in console. Creates .vdjca files
        in the same directory than the fastq.gz files
}
\description{
execute a list of commmand on all .fastq.gz files in `input_dir`
}
\author{
DEBOT Damien <damien.debot@selor.be>
}
