% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_path_program}
\alias{check_path_program}
\title{check_path_program}
\usage{
check_path_program(path, program)
}
\arguments{
\item{path}{the emplacement of the program on the disk (directory or fullpath)}

\item{program}{the name of the program. Can be "mixcr" or "vdjtools"}
}
\value{
the path to the program
}
\description{
check if the path to a program (.jar) is valid
}
\author{
DEBOT Damien <damien.debot@selor.be>
}
