% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{convert_decimal}
\alias{convert_decimal}
\title{convert_decimal}
\usage{
convert_decimal(path, decimal = ",")
}
\arguments{
\item{path}{the path of a .txt file}

\item{decimal}{the decilmark you want to use (default= ,)}
}
\value{
same .txt files with a new decimal separator
}
\description{
change decimalmark to `,` in .txt files.
}
\author{
DEBOT Damien <damien.debot@selor.be>
}
