% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{correct_paths}
\alias{correct_paths}
\title{correct_paths}
\usage{
correct_paths(x)
}
\arguments{
\item{x}{a character string}
}
\value{
the string adapted
}
\description{
change '\' into '/' and add '/' at the end if missing.
}
\author{
DEBOT Damien <damien.debot@selor.be>
}
