% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{txt_to_excel}
\alias{txt_to_excel}
\title{txt_to_excel}
\usage{
txt_to_excel(path, subfolder = NULL)
}
\arguments{
\item{path}{a directory containing .txt files}

\item{subfolder}{a sub-directory of path where you want to save the files}
}
\value{
.xlsx files
}
\description{
convert .txt file to .xlsx file and change decimalmark to `,`.
}
\author{
DEBOT Damien <damien.debot@selor.be>
}
